sync

+++ Code/Libraries/Drivetrain/Drivetrain.cpp
@@ -66,7 +66,7 @@ Drivetrain::~Drivetrain()
 void Drivetrain::goToFishPID(Block block, unsigned long currentTime)
 {
 	//Determine PID output
-	unsigned long dt = currentTime - _previousTime; //Find how long has passed since the last adjustment.
+	 int dt = currentTime - _previousTime; //Find how long has passed since the last adjustment.
 	_previousTime = currentTime;
 
 	Serial.print("dt: ");
